+++ OsmAnd/src/net/osmand/plus/routing/RouteDirectionInfo.java
@@ -101,7 +101,7 @@ public class RouteDirectionInfo {
 
 	// expected time after route point
 	public int getExpectedTime() {
-		return (int) Math.round(distance / averageSpeed);
+		return Math.round(distance / averageSpeed);
 	}
